pyproject.toml
[tool.poetry]
name = "s2-admin-panel"
version = "0.1.0"
description = ""
authors = ["jlllk <djapme@gmail.com>"]
readme = "README.md"
packages = [{include = "s2_admin_panel"}]

[tool.poetry.dependencies]
python = "^3.11"
flake8 = "^6.0.0"
isort = "^5.10.1"
pytest = "^7.2.0"
django = "4.0.4"
uwsgi = "^2.0.21"
django-split-settings = "1.2.0"
python-dotenv = "0.21.0"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"